Readme update
added in some more usefull info in the readme so it is easer to set up on a server. i still need to add in more of my notes.

created a command to test if the python scripts are installed, may extend it to installing the needed libiries if they are missing.